% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayCorrect.R
\name{AssayCorrect}
\alias{AssayCorrect}
\title{AssayCorrect}
\usage{
AssayCorrect(obj, method, nv = 3, npc = 4)
}
\arguments{
\item{obj}{raligner object}

\item{method}{correction method: combat, svd, cPCA}

\item{nv}{represent the correct strength when chosing the svd method}

\item{npc}{represent the correct strength when chosing the cPCA method}
}
\value{
raligner object
}
\description{
AssayCorrect
}
\examples{
ralginer = AssayCorrect(raligner, method = "combat")
}
