% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneStat.R
\name{CalGeneStation}
\alias{CalGeneStation}
\title{Calculate gene station}
\usage{
CalGeneStation(obj, ndims = 70)
}
\arguments{
\item{obj}{raligner object}
}
\value{
gene station information
}
\description{
Calculate gene station
}
\examples{
raligner = CalGeneStation(raligner)

}
