% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateObj.R
\name{createObj}
\alias{createObj}
\title{create raligner object}
\usage{
createObj(
  cell_dat,
  cell_ann,
  bulk_dat,
  bulk_ann,
  comb_ann,
  project_name = "TumorCell"
)
}
\arguments{
\item{cell_dat}{cell line expression information}

\item{cell_ann}{cell line meta information}

\item{bulk_dat}{tumor expression information}

\item{bulk_ann}{tumor meta information}

\item{comb_ann}{combined annotaion information}

\item{project_name}{name of project. eg: tcga_cell}
}
\value{
raligner object
}
\description{
create raligner object
}
\examples{
ralign = createObj(raw_dir = "data/Demo1_tcga_cell/rawdata.rds",
                  project_name = "tcga_ccle",
                  default = "raw")
}
