% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSeuObj.R
\name{CreatSeuObj}
\alias{CreatSeuObj}
\title{Create Seurat object and run dimension reduction}
\usage{
CreatSeuObj(exp_mat, ann, type, ndims = 70)
}
\arguments{
\item{exp_mat}{expression dataframe}

\item{ann}{annotation data}

\item{type}{source of the expression data: bulk/cell}

\item{ndims}{number of principle components}
}
\value{
seurat object
}
\description{
Create Seurat object and run dimension reduction
}
\examples{

seu = CreatSeuObj(exp_mat = exp_mat, ann = ann, type = type)
}
