% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisHelper.R
\name{DegGsea}
\alias{DegGsea}
\title{Analysis gene expression change after correct method of Raligner}
\usage{
DegGsea(obj, genesets = "KEGG")
}
\arguments{
\item{obj}{raligner object after correct analysis}

\item{gensets}{"KEGG", "GO", "Reactome", "Hallmark"}
}
\value{
gsea result
}
\description{
Analysis gene expression change after correct method of Raligner
}
