% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityPlot.R
\name{DenPlot}
\alias{DenPlot}
\title{Title}
\usage{
DenPlot(obj, assay = "raw", top_DE_genes = 1000, mnn = FALSE)
}
\arguments{
\item{obj}{raligner object}

\item{assay}{raw or corrected assay}

\item{top_DE_genes}{top genes to use}
}
\value{
density plot
}
\description{
Title
}
\examples{

DenPlot(obj = ralign, assay = "raw")
}
