% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimension_Reduction.R
\name{DimReduce}
\alias{DimReduce}
\title{Using seurat method for dimension reduction}
\usage{
DimReduce(obj, assay = "raw")
}
\arguments{
\item{obj}{raligner object}

\item{assay}{"raw", "correct"}
}
\value{
raligner object
}
\description{
Using seurat method for dimension reduction
}
\examples{
ralign = DimReduce(ralign, assay = "raw")
}
