% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNN.R
\name{run_MNN}
\alias{run_MNN}
\title{MNN correction}
\usage{
run_MNN(obj, assay = "raw", top_DE_genes = 1000, k1 = 3, k2 = 10, ndist = 3)
}
\arguments{
\item{obj}{raligner object}

\item{assay}{raw or correct}

\item{top_DE_genes}{top differential genes to use}

\item{k1}{knn nubmer of tumors}

\item{k2}{knn number of cell lines}

\item{ndist}{number of knn}
}
\value{
raligner object with MNN correction result
}
\description{
MNN correction
}
\examples{
raligner = run_MNN(raligner)

}
