% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisHelper.R
\name{stroImm}
\alias{stroImm}
\title{Analysis stromal and immune score change after correct method of Raligner}
\usage{
stroImm(obj)
}
\arguments{
\item{obj}{raligner objcet}
}
\value{
stromal and immune score of raw data and correct data
}
\description{
Analysis stromal and immune score change after correct method of Raligner
}
