% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createObj.R
\name{createObj}
\alias{createObj}
\title{create raligner object}
\usage{
createObj(raw_dir, project_name = "TumorCell")
}
\arguments{
\item{raw_dir}{directory of raw data}

\item{project_name}{name of project. eg: tcga_cell}

\item{default}{default assays: raw、combat、cPCA、svd}
}
\value{
raligner object
}
\description{
create raligner object
}
\examples{
ralign = createObj(raw_dir = "data/Demo1_tcga_cell/rawdata.rds",
                  project_name = "tcga_ccle",
                  default = "raw")
}
