% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimPlot.R
\name{DimPlot}
\alias{DimPlot}
\title{Visualization of dimension reduction result}
\usage{
DimPlot(obj, assay = "raw", method = "umap", type = NULL)
}
\arguments{
\item{obj}{raligner object}

\item{method}{umap, tnse}
}
\value{
ggplot result
}
\description{
Visualization of dimension reduction result
}
\examples{
DimPlot(ralign, method = "umap")
}
